refactor: centralize QA utilities
